grand refactoring of NSI
